Oops, huge changes.
Create sub-projects
the first is workerbee-web which is the original code based project.
the second is workerbee-server. now workerbee is running on jetty
embeded server.
the third is workerbee-disto, which assembly other projects to
distribution version.

Another change is directory structure.
add bin and conf directories.

/bin contains executable scripts to enable workerbee start and stop
/conf contains site-specific configuration files, things like
blahblah-site.xml like what you see in hadoop. 

README.txt is added. but the content is empty.
I'll fill it out later.

ok, this commit changes a lot.
I want you NOT to confuse much.